refactor: modularize config flow with helper classes
- Extract entity discovery logic into EntityDiscoveryHelper
- Extract UI formatting logic into UIFormattingHelper
- Split complex ConfigFlow into focused helper modules
- Improve code maintainability and testability
- Update CLAUDE.md with mentorship guidelines
- Fix line length issues in config_flow.py
