+++ custom_components/dali_center/manifest.json
@@ -13,7 +13,7 @@
   "issue_tracker": "https://github.com/maginawin/ha-dali-center/issues",
   "quality_scale": "bronze",
   "requirements": [
-    "PySrDaliGateway==0.4.0"
+    "PySrDaliGateway==0.5.0"
   ],
   "ssdp": [],
   "version": "0.3.0",
