+++ tests/test_config_flow.py
@@ -10,11 +10,9 @@ from homeassistant.config_entries import ConfigEntry
 from custom_components.dali_center.config_flow import (
     OptionsFlowHandler,
     DaliCenterConfigFlow,
-    EntityDiscoveryMixin,
     OPTIONS_SCHEMA
 )
 from custom_components.dali_center.const import DOMAIN
-from custom_components.dali_center.config_flow import dr, er
 from tests.conftest import (
     MockDaliGateway,
     MockDaliGatewayDiscovery,
@@ -57,86 +55,6 @@ class TestConfigFlowConstants:
         assert result["refresh_devices"] is True
         assert result["refresh_groups"] is False  # default
         assert result["refresh_scenes"] is False  # default
-
-
-class TestEntityDiscoveryMixin:
-    """Test EntityDiscoveryMixin functionality."""
-
-    @pytest.fixture
-    def mixin_instance(self):
-        """Create EntityDiscoveryMixin instance for testing."""
-        return EntityDiscoveryMixin()
-
-    @pytest.mark.asyncio
-    async def test_discover_entities_all_types(self, mixin_instance):
-        """Test discovering all entity types."""
-        mock_gateway = MockDaliGateway()
-
-        # Mock the discovery methods that would be called
-        with patch.object(
-            mock_gateway, "discover_devices", new_callable=AsyncMock
-        ) as mock_discover_devices, \
-            patch.object(
-            mock_gateway, "discover_groups", new_callable=AsyncMock
-        ) as mock_discover_groups, \
-            patch.object(
-                mock_gateway, "discover_scenes", new_callable=AsyncMock
-        ) as mock_discover_scenes:
-
-            # Set up mock return values
-            mock_discover_devices.return_value = mock_gateway.devices
-            mock_discover_groups.return_value = mock_gateway.groups
-            mock_discover_scenes.return_value = mock_gateway.scenes
-
-            result = await mixin_instance._discover_entities(mock_gateway)  # pylint: disable=protected-access
-
-            # Verify all discovery methods were called
-            mock_discover_devices.assert_called_once()
-            mock_discover_groups.assert_called_once()
-            mock_discover_scenes.assert_called_once()
-
-            # Verify results
-            assert "devices" in result
-            assert "groups" in result
-            assert "scenes" in result
-            assert result["devices"] == mock_gateway.devices
-            assert result["groups"] == mock_gateway.groups
-            assert result["scenes"] == mock_gateway.scenes
-
-    @pytest.mark.asyncio
-    async def test_discover_entities_devices_only(self, mixin_instance):
-        """Test discovering only devices."""
-        mock_gateway = MockDaliGateway()
-
-        with patch.object(
-            mock_gateway, "discover_devices", new_callable=AsyncMock
-        ) as mock_discover_devices, \
-            patch.object(
-                mock_gateway, "discover_groups", new_callable=AsyncMock
-        ) as mock_discover_groups, \
-            patch.object(
-                mock_gateway, "discover_scenes", new_callable=AsyncMock
-        ) as mock_discover_scenes:
-
-            mock_discover_devices.return_value = mock_gateway.devices
-
-            result = await mixin_instance._discover_entities(  # pylint: disable=protected-access
-                mock_gateway,
-                discover_devices=True,
-                discover_groups=False,
-                discover_scenes=False
-            )
-
-            # Verify only devices were discovered
-            mock_discover_devices.assert_called_once()
-            mock_discover_groups.assert_not_called()
-            mock_discover_scenes.assert_not_called()
-
-            # Verify results
-            assert "devices" in result
-            assert "groups" not in result
-            assert "scenes" not in result
-            assert result["devices"] == mock_gateway.devices
 
 
 class TestDaliCenterConfigFlow:
@@ -279,108 +197,6 @@ class TestDaliCenterConfigFlowComplete:
                 mock_configure.assert_called_once()
                 assert config_flow._selected_gateway is not None
 
-    @pytest.mark.asyncio
-    async def test_async_step_configure_entities_with_user_input(
-            self, config_flow):
-        """Test configure entities with user selection."""
-        mock_gateway = MockDaliGateway()
-        config_flow._selected_gateway = mock_gateway
-        config_flow._discovered_entities = {
-            "devices": [{"sn": "dev1", "name": "Device 1"}],
-            "groups": [],
-            "scenes": []
-        }
-
-        user_input = {"devices": ["dev1"]}
-
-        with patch.object(
-            config_flow, "_filter_selected_entities"
-        ) as mock_filter:
-            mock_filter.return_value = {"devices": [
-                {"sn": "dev1", "name": "Device 1"}]}
-
-            with patch.object(config_flow, "async_create_entry") as mock_create:
-                mock_create.return_value = {
-                    "type": FlowResultType.CREATE_ENTRY}
-
-                await config_flow.async_step_configure_entities(user_input)
-
-                mock_create.assert_called_once()
-
-    def test_prepare_entity_selection_schema_with_devices(self, config_flow):
-        """Test preparing entity selection schema with devices."""
-        devices = [
-            {"sn": "dev1", "name": "Device 1",
-                "dev_type": "light", "unique_id": "gw123_dev1"},
-            {"sn": "dev2", "name": "Device 2",
-                "dev_type": "sensor", "unique_id": "gw123_dev2"}
-        ]
-        groups = []
-        scenes = []
-
-        schema = config_flow._prepare_entity_selection_schema(
-            devices, groups, scenes)
-
-        assert schema is not None
-
-    def test_prepare_entity_selection_schema_with_groups(self, config_flow):
-        """Test preparing entity selection schema with groups."""
-        devices = []
-        groups = [
-            {"sn": "grp1", "name": "Group 1",
-                "unique_id": "gw123_grp1", "channel": 1, "id": 1},
-            {"sn": "grp2", "name": "Group 2",
-                "unique_id": "gw123_grp2", "channel": 1, "id": 2}
-        ]
-        scenes = []
-
-        schema = config_flow._prepare_entity_selection_schema(
-            devices, groups, scenes)
-
-        assert schema is not None
-
-    def test_prepare_entity_selection_schema_with_scenes(self, config_flow):
-        """Test preparing entity selection schema with scenes."""
-        devices = []
-        groups = []
-        scenes = [
-            {"sn": "scn1", "name": "Scene 1",
-                "unique_id": "gw123_scn1", "channel": 1, "id": 1},
-            {"sn": "scn2", "name": "Scene 2",
-                "unique_id": "gw123_scn2", "channel": 1, "id": 2}
-        ]
-
-        schema = config_flow._prepare_entity_selection_schema(
-            devices, groups, scenes)
-
-        assert schema is not None
-
-    def test_filter_selected_entities_all_types(self, config_flow):
-        """Test filtering selected entities for all types."""
-        user_input = {
-            "devices": ["gw123_dev1"],
-            "groups": ["gw123_grp1"],
-            "scenes": ["gw123_scn1"]
-        }
-        discovered_entities = {
-            "devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"}
-            ],
-            "groups": [
-                {"sn": "grp1", "name": "Group 1", "unique_id": "gw123_grp1"}
-            ],
-            "scenes": [
-                {"sn": "scn1", "name": "Scene 1", "unique_id": "gw123_scn1"}
-            ]
-        }
-
-        result = config_flow._filter_selected_entities(
-            user_input, discovered_entities)
-
-        assert len(result["devices"]) == 1
-        assert len(result["groups"]) == 1
-        assert len(result["scenes"]) == 1
-
 
 class TestOptionsFlowHandlerComplete:
     """Test additional OptionsFlowHandler functionality."""
@@ -416,91 +232,6 @@ class TestOptionsFlowHandlerComplete:
         return flow
 
     @pytest.mark.asyncio
-    async def test_async_step_refresh_with_entity_discovery(self, options_flow):
-        """Test refresh step with entity discovery."""
-        options_flow._refresh_devices = True
-
-        # Mock runtime_data to avoid gateway_not_found abort
-        mock_runtime_data = Mock()
-        mock_gateway = MockDaliGateway()
-        mock_runtime_data.gateway = mock_gateway
-        options_flow._config_entry.runtime_data = mock_runtime_data
-
-        with patch.object(options_flow, "_discover_entities") as mock_discover:
-            mock_discover.return_value = {
-                "devices": [{"sn": "dev2", "name": "Device 2"}],
-                "groups": [],
-                "scenes": []
-            }
-
-            with patch.object(
-                options_flow, "async_step_select_entities"
-            ) as mock_select:
-                mock_select.return_value = {"type": FlowResultType.FORM}
-
-                await options_flow.async_step_refresh()
-
-                mock_select.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_async_step_select_entities_with_user_input(
-            self, options_flow):
-        """Test select entities step with user selection."""
-        options_flow._discovered_entities = {
-            "devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"}
-            ],
-            "groups": [],
-            "scenes": []
-        }
-
-        user_input = {"devices": ["gw123_dev1"]}
-
-        with patch.object(
-            options_flow, "_filter_selected_entities"
-        ) as mock_filter:
-            mock_filter.return_value = {"devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"}]}
-
-            with patch.object(
-                options_flow, "async_step_refresh_result"
-            ) as mock_result:
-                mock_result.return_value = {
-                    "type": FlowResultType.CREATE_ENTRY}
-
-                # Mock the necessary registry operations
-                with patch(f"{CFM}.dr.async_get") as mock_dr, \
-                        patch(f"{CFM}.er.async_get") as mock_er:
-                    mock_device_reg = Mock()
-                    mock_entity_reg = Mock()
-                    mock_dr.return_value = mock_device_reg
-                    mock_er.return_value = mock_entity_reg
-                    mock_device_reg.async_remove_device = Mock()
-                    mock_entity_reg.async_remove = Mock()
-
-                    # Mock registry entries
-                    with patch.object(
-                        dr, "async_entries_for_config_entry", return_value=[]
-                    ), patch.object(
-                        er, "async_entries_for_config_entry", return_value=[]
-                    ):
-
-                        # Mock hass.config_entries.async_update_entry to avoid
-                        # the data modification error
-                        options_flow.hass.config_entries.async_update_entry = (
-                            Mock()
-                        )
-
-                        # Mock hass.async_create_task for the reload task
-                        options_flow.hass.async_create_task = Mock()
-
-                        await options_flow.async_step_select_entities(
-                            user_input
-                        )
-
-                        mock_result.assert_called_once()
-
-    @pytest.mark.asyncio
     async def test_async_step_refresh_result_with_changes(self, options_flow):
         """Test refresh result step with entity changes."""
         options_flow._refresh_results = {
@@ -519,54 +250,3 @@ class TestOptionsFlowHandlerComplete:
             await options_flow.async_step_refresh_result(user_input={})
 
             mock_create.assert_called_once()
-
-    def test_calculate_entity_differences_added_devices(self, options_flow):
-        """Test calculating entity differences with added devices."""
-        # Set refresh flag to enable devices processing
-        options_flow._refresh_devices = True
-
-        selected = {
-            "devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"},
-                {"sn": "dev2", "name": "Device 2", "unique_id": "gw123_dev2"}
-            ],
-            "groups": [],
-            "scenes": []
-        }
-        current_data = {
-            "devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"}
-            ],
-            "groups": [],
-            "scenes": []
-        }
-
-        options_flow._calculate_entity_differences(selected, current_data)
-
-        assert len(options_flow._refresh_results["devices_added"]) == 1
-        assert options_flow._refresh_results["devices_added"][0]["sn"] == "dev2"
-
-    def test_calculate_entity_differences_removed_devices(self, options_flow):
-        """Test calculating entity differences with removed devices."""
-        # Set refresh flag to enable devices processing
-        options_flow._refresh_devices = True
-
-        selected = {
-            "devices": [],
-            "groups": [],
-            "scenes": []
-        }
-        current_data = {
-            "devices": [
-                {"sn": "dev1", "name": "Device 1", "unique_id": "gw123_dev1"}
-            ],
-            "groups": [],
-            "scenes": []
-        }
-
-        options_flow._calculate_entity_differences(selected, current_data)
-
-        assert len(options_flow._refresh_results["devices_removed"]) == 1
-        assert (
-            options_flow._refresh_results["devices_removed"][0]["sn"] == "dev1"
-        )
